delete book using local method

+++ src/controllers/book.ts
@@ -496,26 +496,29 @@ export const deleteBook: RequestHandler = async (req, res) => {
     return res.json({ success: false });
   }
 
+  // remove old book file (epub) from storage
+  const uploadPath = path.join(__dirname, "../books");
+  const oldFilePath = path.join(uploadPath, book.fileInfo.id);
+
+  if (!fs.existsSync(oldFilePath))
+    return sendErrorResponse({
+      message: "Book file not found!",
+      status: 404,
+      res,
+    });
+
+  fs.unlinkSync(oldFilePath);
+
+  if (book.cover?.id) {
+    await cloudinary.uploader.destroy(book.cover.id);
+  }
+
   await BookModel.findByIdAndDelete(book._id);
   const author = await AuthorModel.findById(user.authorId);
   if (author) {
     author.books = author.books.filter((id) => id.toString() !== bookId);
     await author.save();
   }
-
-  if (book.cover?.id) {
-    const coverDeleteCommand = new DeleteObjectCommand({
-      Bucket: process.env.AWS_PUBLIC_BUCKET,
-      Key: book.cover.id,
-    });
-    await s3Client.send(coverDeleteCommand);
-  }
-
-  const bookFileDeleteCommand = new DeleteObjectCommand({
-    Bucket: process.env.AWS_PRIVATE_BUCKET,
-    Key: book.fileInfo.id,
-  });
-  await s3Client.send(bookFileDeleteCommand);
 
   res.json({ success: true });
 };
